exit out if the propagation script can't download the source

+++ prop.sh
@@ -4,6 +4,10 @@ if [ ! -f QueryLogging.txt ]; then
     touch QueryLogging.txt
 fi
 wget http://bryantlab02.rtp.raleigh.ibm.com/public/master.zip
+if [ $? -ne 0 ];then
+    echo "problem downloading the master.zip exiting"
+    exit -1
+fi
 rm -rf /var/www/htdocs/selfservice/
 mkdir -p /var/www/htdocs/selfservice/
 unzip -o master.zip -d /var/www/htdocs/selfservice/
